Move components to addon and re-export in app to follow addon guidelines
